@@ -17,8 +17,11 @@
 package com.ctrip.framework.apollo.spring.property;
 
 import com.ctrip.framework.apollo.ConfigChangeListener;
+import com.ctrip.framework.apollo.build.ApolloInjector;
 import com.ctrip.framework.apollo.model.ConfigChangeEvent;
+import com.ctrip.framework.apollo.spring.events.ApolloConfigChangeEvent;
 import com.ctrip.framework.apollo.spring.util.SpringInjector;
+import com.ctrip.framework.apollo.util.ConfigUtil;
 import com.google.gson.Gson;
 import java.lang.reflect.Field;
 import java.lang.reflect.Type;
@@ -26,34 +29,38 @@ import java.util.Collection;
 import java.util.Set;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.beans.BeansException;
 import org.springframework.beans.TypeConverter;
 import org.springframework.beans.factory.config.ConfigurableBeanFactory;
-import org.springframework.beans.factory.config.ConfigurableListableBeanFactory;
-import org.springframework.core.env.Environment;
+import org.springframework.context.ApplicationContext;
+import org.springframework.context.ApplicationContextAware;
+import org.springframework.context.ApplicationListener;
+import org.springframework.context.ConfigurableApplicationContext;
 import org.springframework.util.CollectionUtils;
 
 /**
  * Create by zhangzheng on 2018/3/6
  */
-public class AutoUpdateConfigChangeListener implements ConfigChangeListener{
-  private static final Logger logger = LoggerFactory.getLogger(AutoUpdateConfigChangeListener.class);
+public class AutoUpdateConfigChangeListener implements ConfigChangeListener,
+    ApplicationListener<ApolloConfigChangeEvent>, ApplicationContextAware {
+
+  private static final Logger logger = LoggerFactory.getLogger(
+      AutoUpdateConfigChangeListener.class);
 
   private final boolean typeConverterHasConvertIfNecessaryWithFieldParameter;
-  private final Environment environment;
-  private final ConfigurableBeanFactory beanFactory;
-  private final TypeConverter typeConverter;
+  private ConfigurableBeanFactory beanFactory;
+  private TypeConverter typeConverter;
   private final PlaceholderHelper placeholderHelper;
   private final SpringValueRegistry springValueRegistry;
   private final Gson gson;
+  private final ConfigUtil configUtil;
 
-  public AutoUpdateConfigChangeListener(Environment environment, ConfigurableListableBeanFactory beanFactory){
+  public AutoUpdateConfigChangeListener() {
     this.typeConverterHasConvertIfNecessaryWithFieldParameter = testTypeConverterHasConvertIfNecessaryWithFieldParameter();
-    this.beanFactory = beanFactory;
-    this.typeConverter = this.beanFactory.getTypeConverter();
-    this.environment = environment;
     this.placeholderHelper = SpringInjector.getInstance(PlaceholderHelper.class);
     this.springValueRegistry = SpringInjector.getInstance(SpringValueRegistry.class);
     this.gson = new Gson();
+    this.configUtil = ApolloInjector.getInstance(ConfigUtil.class);
   }
 
   @Override
@@ -90,7 +97,9 @@ public class AutoUpdateConfigChangeListener implements ConfigChangeListener{
 
   /**
    * Logic transplanted from DefaultListableBeanFactory
-   * @see org.springframework.beans.factory.support.DefaultListableBeanFactory#doResolveDependency(org.springframework.beans.factory.config.DependencyDescriptor, java.lang.String, java.util.Set, org.springframework.beans.TypeConverter)
+   *
+   * @see org.springframework.beans.factory.support.DefaultListableBeanFactory#doResolveDependency(org.springframework.beans.factory.config.DependencyDescriptor,
+   * java.lang.String, java.util.Set, org.springframework.beans.TypeConverter)
    */
   private Object resolvePropertyValue(SpringValue springValue) {
     // value will never be null, as @Value and @ApolloJsonValue will not allow that
@@ -98,7 +107,7 @@ public class AutoUpdateConfigChangeListener implements ConfigChangeListener{
         .resolvePropertyValue(beanFactory, springValue.getBeanName(), springValue.getPlaceholder());
 
     if (springValue.isJson()) {
-      value = parseJsonValue((String)value, springValue.getGenericType());
+      value = parseJsonValue((String) value, springValue.getGenericType());
     } else {
       if (springValue.isField()) {
         // org.springframework.beans.TypeConverter#convertIfNecessary(java.lang.Object, java.lang.Class, java.lang.reflect.Field) is available from Spring 3.2.0+
@@ -134,5 +143,20 @@ public class AutoUpdateConfigChangeListener implements ConfigChangeListener{
     }
 
     return true;
+  }
+
+  @Override
+  public void setApplicationContext(ApplicationContext applicationContext) throws BeansException {
+    //it is safe enough to cast as all known application context is derived from ConfigurableApplicationContext
+    this.beanFactory = ((ConfigurableApplicationContext) applicationContext).getBeanFactory();
+    this.typeConverter = this.beanFactory.getTypeConverter();
+  }
+
+  @Override
+  public void onApplicationEvent(ApolloConfigChangeEvent event) {
+    if (!configUtil.isAutoUpdateInjectedSpringPropertiesEnabled()) {
+      return;
+    }
+    this.onChange(event.getConfigChangeEvent());
   }
 }
