@@ -30,7 +30,7 @@ import javax.persistence.Table;
  */
 @Entity
 @Table(name = "RolePermission")
-@SQLDelete(sql = "Update RolePermission set isDeleted = 1 where id = ?")
+@SQLDelete(sql = "Update RolePermission set IsDeleted = 1, DeletedAt = ROUND(UNIX_TIMESTAMP(NOW(4))*1000) where Id = ?")
 @Where(clause = "isDeleted = 0")
 public class RolePermission extends BaseEntity {
   @Column(name = "RoleId", nullable = false)
