@@ -30,7 +30,7 @@ import javax.persistence.Table;
  */
 @Entity
 @Table(name = "UserRole")
-@SQLDelete(sql = "Update UserRole set isDeleted = 1 where id = ?")
+@SQLDelete(sql = "Update UserRole set IsDeleted = 1, DeletedAt = ROUND(UNIX_TIMESTAMP(NOW(4))*1000) where Id = ?")
 @Where(clause = "isDeleted = 0")
 public class UserRole extends BaseEntity {
   @Column(name = "UserId", nullable = false)
